Fixes and add unit test
