+++ src/applications/facility-locator/tests/helpers/radiusFromBoundingBox.unit.spec.js
@@ -1,0 +1,30 @@
+import { expect } from 'chai';
+import { radiusFromBoundingBox } from '../../utils/facilityDistance';
+import { MIN_RADIUS } from '../../constants';
+
+describe('radiusFromBoundingBox', () => {
+  it('should return a valid computed radius - Dallas', () => {
+    const testDallasFeatureBbox = [
+      {
+        bbox: [
+          -97.04103101,
+          32.6185720016946,
+          -96.5555139885032,
+          33.016494008014,
+        ],
+      },
+    ];
+    const testRad = radiusFromBoundingBox(testDallasFeatureBbox);
+    expect(testRad > MIN_RADIUS).to.eql(true);
+  });
+
+  it('should return a default 10 miles radius - zip 92052', () => {
+    const testDallasFeatureBbox = [
+      {
+        bbox: [-117.357412, 33.21172839552, -117.35718939552, 33.21189160448],
+      },
+    ];
+    const testRad = radiusFromBoundingBox(testDallasFeatureBbox);
+    expect(testRad === MIN_RADIUS).to.eql(true);
+  });
+});
